% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_adv}
\alias{check_adv}
\title{checks adv}
\usage{
check_adv(chunk.adv, tests = "all", verbose = FALSE)
}
\arguments{
\item{chunk.adv}{a data.frame created with load.ADV, with the window.idx column}

\item{tests}{a character array of test names, or 'all' to run all tests}

\item{verbose}{boolean for diagnostic print outs}
}
\value{
failed, T or F
}
\description{
checks data for various quality metrics  \cr
}
\details{
a \code{GDopp} function for checking data quality.\cr
}
\examples{
\dontrun{
folder.nm  <- system.file('extdata', package = 'GDopp')
file.nm <- "ALQ102.dat"
data.adv <- load_adv(file.nm=file.nm, folder.nm =folder.nm)
window.adv <- window_adv(data.adv,freq=32,window.mins=10)
chunk.adv <- window.adv[window.adv$window.idx==7, ]
check.adv(chunk.adv,tests=c('signal.noise.check_adv','frozen.turb.check_adv'),verbose=TRUE)
}
}
\references{
Vachon, Dominic, Yves T. Prairie, and Jonathan J. Cole.
\emph{The relationship between near-surface turbulence and gas transfer
velocity in freshwater systems and its implications for floating chamber measurements of gas exchange}.
Limnology and Oceanography 55, no. 4 (2010): 1723.

Kitaigorodskii, S. A., M. A. Donelan, J. L. Lumley, and E. A. Terray.
\emph{Wave-turbulence interactions in the upper ocean. Part II. Statistical
characteristics of wave and turbulent components of the random velocity
field in the marine surface layer}. Journal of Physical Oceanography 13, no. 11 (1983): 1988-1999.

#'Lien, Ren-Chieh, and Eric A. D'Asaro. \emph{Measurement of turbulent kinetic energy dissipation rate with a Lagrangian float.}
Journal of Atmospheric and Oceanic Technology 23, no. 7 (2006): 964-976.
}
\keyword{math}
\keyword{methods,}

