% Generated by roxygen2 (4.0.1): do not edit by hand
\name{epsilon_to_k}
\alias{epsilon_to_k}
\title{convert epsilon to k600}
\usage{
epsilon_to_k(epsilon, temperature = 20, nu = 0.2)
}
\arguments{
\item{epsilon}{Turbulence dissipation rate}

\item{temperature}{Temperature in degC of near surface water}

\item{nu}{Constant of proportionality}
}
\value{
a single k600 value
}
\description{
Uses Zappa et al. 2007 to convert epsilon to k600.  \cr
}
\details{
a \code{GDopp} function for converting epsilon to k600.\cr
}
\examples{
\dontrun{
folder.nm <- system.file('extdata', package = 'GDopp')
file.nm <- "ALQ102.dat"
block.use <- 7
data.adv <- load_adv(file.nm=file.nm, folder.nm =folder.nm)
window.adv <- window_adv(data.adv,freq=32,window.mins=10)
data.sen <- load_sen(paste0(substr(file.nm,start=1,stop=nchar(file.nm)-4),'.sen'))
temp.adv <- temp_calc(data.sen,window.adv$window.idx,freq=32)
chunk.adv <- window.adv[window.adv$window.idx==block.use, ]
epsilon <- fit_epsilon(chunk.adv,freq=32)
epsilon_to_k(epsilon,temperature=temp.adv[block.use],nu=0.2)
}
}
\references{
Zappa, Christopher J., Wade R. McGillis, Peter A. Raymond, James B. Edson, Eric J. Hintsa,
Hendrik J. Zemmelink, John WH Dacey, and David T. Ho.
\emph{Environmental turbulent mixing controls on air/water gas exchange in marine and aquatic systems.}
Geophysical Research Letters 34, no. 10 (2007).
}
\keyword{epsilon2k}

