% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_kin_viscosity}
\alias{get_kin_viscosity}
\title{get kinematic viscosity}
\usage{
get_kin_viscosity(temperature = 20)
}
\arguments{
\item{temperature}{Temperature in degC of water}
}
\value{
kinematic viscosity
}
\description{
get kinematic viscosity
}
\examples{
kin_vis <- get_kin_viscosity(temperature = 20)
}
\keyword{kinematic.vis}

