% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_metadata}
\alias{get_metadata}
\title{get measurement details}
\usage{
get_metadata(deploy_nm, folder_nm)
}
\arguments{
\item{deploy_nm}{a valid deployment name (string)}

\item{folder_nm}{the folder where the deployment files reside.}
}
\value{
a list of measurement metadata
}
\description{
get measurement details
}
\details{
a \code{GDopp} function for extracting sampling info from file.\cr
}
\examples{
folder_nm <- system.file('extdata', package = 'GDopp')
deploy_nm <- "ALQ102"
get_metadata(deploy_nm, folder_nm)
}
\keyword{methods}

