\name{kinematic.vis}
\alias{kinematic.vis}
\title{get kinematic viscosity}
\usage{
  kinematic.vis(temperature = 20)
}
\arguments{
  \item{\code{temperature}}{Temperature in °C of water}
}
\value{
  kinematic viscosity
}
\description{
  get kinematic viscosity
}
\examples{
kin.vis <- kinematic.vis(temperature=20)
}
\keyword{kinematic.vis}

