% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kinematic_vis}
\alias{kinematic_vis}
\title{get kinematic viscosity}
\usage{
kinematic_vis(temperature = 20)
}
\arguments{
\item{temperature}{Temperature in degC of water}
}
\value{
kinematic viscosity
}
\description{
get kinematic viscosity
}
\examples{
kin.vis <- kinematic_vis(temperature=20)
}
\keyword{kinematic.vis}

