% Generated by roxygen2 (4.0.1): do not edit by hand
\name{temp.calc}
\alias{temp.calc}
\title{block-averaged temperature}
\usage{
temp.calc(data.sen, window.idx, freq = 32, calc.time = FALSE)
}
\arguments{
\item{data.sens}{a data.frame created with load.ADV, with the window.idx column}

\item{window.idx}{window.idx column from adv data.frame.}

\item{freq}{the frequency (Hz) of the measurements}

\item{calc.time}{a boolean for including a time vector in the output (will then return a data.frame)}
}
\value{
a vector of averaged values
}
\description{
Water temperature average for within block  \cr
}
\details{
a \code{GDopp} function for averaging water temperature measurements from ADV sensor (1Hz).\cr
assumes one second spin-up for high-frequency measurement, and a likely incomplete final second for high frequency.
}
\examples{
\dontrun{
folder.nm <- system.file('extdata', package = 'GDopp')
file.nm <- "ALQ102.dat"
data.adv <- load.ADV(file.nm=file.nm, folder.nm =folder.nm)
window.adv <- window_ADV(data.adv,freq=32,window.mins=10)
data.sen <- load.sen(file.nm="ALQ102.sen", folder.nm = folder.nm)

temp.calc(data.sen,window.adv$window.idx,freq=32)
}
}
\keyword{math}
\keyword{methods,}

