% Generated by roxygen2 (4.0.1): do not edit by hand
\name{temp_calc}
\alias{temp_calc}
\title{block-averaged temperature}
\usage{
temp_calc(data.sen, window.idx, freq = 32)
}
\arguments{
\item{data.sen}{a data.frame created with load.ADV, with the window.idx column}

\item{window.idx}{window.idx column from adv data.frame.}

\item{freq}{the frequency (Hz) of the measurements}
}
\value{
a vector of averaged values
}
\description{
Water temperature average for within block  \cr
}
\details{
a \code{GDopp} function for averaging water temperature measurements from ADV sensor (1Hz).\cr
assumes one second spin-up for high-frequency measurement, and a likely incomplete final second for high frequency.
}
\examples{
\dontrun{
folder.nm <- system.file('extdata', package = 'GDopp')
file.nm <- "ALQ102.dat"
data.adv <- load_adv(file.nm=file.nm, folder.nm =folder.nm)
window.adv <- window_adv(data.adv,freq=32,window.mins=10)
data.sen <- load_sen(file.nm="ALQ102.sen", folder.nm = folder.nm)

temps <- temp_calc(data.sen,window.adv$window.idx,freq=32)
time <- time_calc(data.sen,window.adv$window.idx,freq=32)
plot(time,temps)
}
}
\seealso{
\link{time_calc}, \link{load_adv}, \link{load_sen}
}
\keyword{math}
\keyword{methods,}

