% Generated by roxygen2 (4.0.1): do not edit by hand
\name{velocity_calc}
\alias{velocity_calc}
\title{gets along-stream mean velocity}
\usage{
velocity_calc(chunk.adv)
}
\arguments{
\item{chunk.adv}{a data.frame created with load.ADV, with the window.idx column}
}
\value{
a single mean along-stream velocity value
}
\description{
gets along-stream mean velocity  \cr
}
\details{
a \code{GDopp} function for calculating along-stream mean velocity.\cr
}
\examples{
\dontrun{
folder.nm <- system.file('extdata', package = 'GDopp')
file.nm <- "ALQ102.dat"
data.adv <- load_adv(file.nm=file.nm, folder.nm =folder.nm)
window.adv <- window_adv(data.adv,freq=32,window.mins=10)
chunk.adv <- window.adv[window.adv$window.idx==7, ]
velocity_calc(chunk.adv)
}
}
\keyword{v.calc}

