% Generated by roxygen2 (4.0.1): do not edit by hand
\name{window_coord}
\alias{window_coord}
\title{block-averaged coordinate shift}
\usage{
window_coord(data.sen, window.idx, freq = 32)
}
\arguments{
\item{data.sen}{a data.frame created with load.ADV, with the window.idx column}

\item{window.idx}{window.idx column from adv data.frame.}

\item{freq}{the frequency (Hz) of the measurements}
}
\description{
Coord system components averaged within a block  \cr
}
\details{
a \code{GDopp} function for averaging coordinate measurements from ADV sensor (1Hz).\cr
assumes one second spin-up for high-frequency measurement, and a likely incomplete final second for high frequency.
}
\examples{
\dontrun{
folder.nm <- system.file('extdata', package = 'GDopp')
file.nm <- "ALQ102.dat"
data.adv <- load.ADV(file.nm=file.nm, folder.nm =folder.nm)
window.adv <- window_ADV(data.adv,freq=32,window.mins=10)
data.sen <- load.sen(file.nm="ALQ102.sen", folder.nm = folder.nm)

window_coord(data.sen,window.adv$window.idx,freq=32)
}
}
\keyword{math}
\keyword{methods,}

