% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coord_transform}
\alias{coord_transform}
\title{coordinate transformation}
\usage{
coord_transform(trans_matrix, data_v, position_data)
}
\arguments{
\item{trans_matrix}{a matrix of tranformation values (static) in 3 x 3 matrix.}
}
\description{
coord_transform is modified from Transform.m (from http://www.nortek-
as.com/en/knowledge-center), which is a Matlab script that shows how
velocity data can be transformed between beam coordinates and ENU coordinates.
Beam coordinates are defined as the velocity measured along the three
beams of the instrument.
ENU coordinates are defined in an earth coordinate system, where
E represents the East-West component, N represents the North-South
component and U represents the Up-Down component.

Note that the transformation matrix must be recalculated every time
the orientation, heading, pitch or roll changes.
}
\references{
Lohrmann, Atle, Ramon Cabrera, and Nicholas C. Kraus. \emph{Acoustic-
Doppler velocimeter (ADV) for laboratory use.} In Fundamentals and
advancements in hydraulic measurements and experimentation, pp. 351-365. ASCE, 1994.
}

