% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kinematic.vis}
\alias{kinematic.vis}
\title{get kinematic viscosity}
\usage{
kinematic.vis(temperature = 20)
}
\arguments{
\item{temperature}{Temperature in degC of water}
}
\value{
kinematic viscosity
}
\description{
get kinematic viscosity
}
\examples{
kin.vis <- kinematic.vis(temperature=20)
}
\keyword{kinematic.vis}

