% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.sen}
\alias{load.sen}
\title{load sensor diagnostic data}
\usage{
load.sen(file.nm,
  folder.nm = "../../../Desktop/Science Projects/GDopp/supporting data/")
}
\arguments{
\item{file.nm}{a valid file name.}

\item{folder.nm}{the folder where the data file resides.}
}
\value{
a GDopp data.frame.
}
\description{
Assumes the following format: \cr
\tabular{rlll}{
1 \tab Month   \tab (1-12) \cr
2 \tab Day  \tab (1-31)\cr
3 \tab Year \tab \cr
4 \tab Hour \tab (0-23)\cr
5 \tab Minute    \tab (0-59)\cr
6 \tab Second  \tab (0-59) \cr
7 \tab Error code  \tab  \cr
8 \tab Status code  \tab  \cr
9 \tab Battery voltage  \tab (V) \cr
10 \tab Speed of sound  \tab (m/s) \cr
11 \tab Heading  \tab (deg) \cr
12 \tab Pitch  \tab (deg) \cr
13 \tab Roll  \tab (deg) \cr
14 \tab Temperature  \tab (degC) \cr
15 \tab Analog input \tab  \cr
16 \tab Checksum \tab (1=failed)\cr
}
}
\details{
a \code{GDopp} function for loading sensor diagnostic data from a file.\cr
}
\examples{
folder.nm <- '../../../Desktop/Science Projects/GDopp/supporting data/'
file.nm <- "ICACOS04.sen"
data.sen <- load.sen(file.nm=file.nm, folder.nm =folder.nm)
}
\keyword{load.sen}

